Add files via upload
Exercice : https://www.codingame.com/ide/puzzle/mime-type
The MIME type is used in many internet protocols to associate a media type (html, image, video, ...) with the content sent. This MIME type is usually deduced from the extension of the file to be transferred.
You need to write a program that can detect the MIME type of a file from its name.